+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2026-03-?? 4.1.43
     - remove product_configuration from EEPROMEditor (merge into model)
     - attempt to dynamically scale height of area scan view
+    - add median, IQR to Prod.PixelNoise and fixed math
 - 2026-03-30 4.1.42
     - track XS laser passwords by SN
     - display laser password status via StatusIndicators